[MOV] website_sale: move payment-specific logic to `website_payment`
The `website_id` field is defined in the `website_payment` but not used
for anything until `website_sale` is installed.

This commits moves the `_get_compatible_providers` override filtering
providers based on the current website from `website_sale` to
`website_payment`, where the `website_id` field is defined.

Task-3084364
